fix: rename interfaces ts, simplified store
