+++ resources/js/pages/dashboard/Props.ts
@@ -1,9 +1,9 @@
-import { chartData } from "../../components/charts/Props";
+import { ChartData } from "../../components/charts/Props";
 
 export interface dashboardState {
     count_example: number,
     data_example: {
         labels: Array<string>,
-        datasets: Array<chartData>
+        datasets: Array<ChartData>
     },
 }
